requirements.txt
Django>=5.2
django-unfold
python-dotenv
django-import-export

scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
jupyterlab>=4.0
ipython>=8.0